feat: import como propostas em aberto, campos distrito/conselho/grupo económico, auto-cálculo EditSale
- Import CSV cria propostas com status 'sent' em vez de vendas
- Dedup de propostas por CPE serial number
- Novos campos no cliente: Distrito (obrigatório), Conselho (obrigatório), Grupo Económico (opcional)
- EditSaleModal: auto-cálculo de margem e comissão para vendas de energia
- findValue com match parcial para headers de Excel com nomes longos
- parseNumericValue corrigido para formato europeu (1.234,56)

+++ src/components/proposals/CreateProposalModal.tsx
@@ -336,7 +336,7 @@ export function CreateProposalModal({ client, open, onOpenChange, onSuccess, pre
                 fidelizacao_end: cpe.fidelizacao_end || null,
                 notes: cpe.notes || null,
                 consumo_anual: cpe.consumo_anual ? parseFloat(cpe.consumo_anual) : null,
-                duracao_contrato: cpe.duracao_contrato ? parseInt(cpe.duracao_contrato) : null,
+                duracao_contrato: cpe.duracao_contrato ? parseFloat(cpe.duracao_contrato) : null,
                 dbl: cpe.dbl ? parseFloat(cpe.dbl) : null,
                 margem: cpe.margem ? parseFloat(cpe.margem) : null,
                 comissao: cpe.comissao ? parseFloat(cpe.comissao) : null,
